requirements.txt
apache-airflow
pytest
pandas
boto3
sqlalchemy
psycopg2-binary
